each team members explanation for team bravo.
Marliegh: 1-21, 45-51
So some context to the issue we did is that we had to figure out the file extension because the issue is we are not getting the right file type. So we had to change it to ABR. We created several variables that get the location of the file, the file path name, file type being saved, file name, the index of the filename start and file type, and length of file name/path. These variables will later help when it comes to the user entering the file name and seeing if it is ABR.  An important variable is myFile; it is the file name and the sample/code below is what the user is typing. From that you can see if the extension is right or wrong.And another important variable is myFileType which is what we are trying to find out. In the lines 45-51 if the file type does not equal ABR it stops. And if it does not pass then it tells the user they selected the wrong file type and to try again. It does this while my file type does not equal ABR. And when it does equal ABR the file load is done.


Shreena: 23-42:
 In this part of the program, there is a do while loop. In the do part of the loop, we are calling the method to get the file name with the path. For example: “C:\Documents\Test.ABR”. it is nameLength - 1 because we are starting from the end because the "."  There is a for loop that is in the do loop. In the for loop, we have the int i = (filenameLength - 1); i >= 0; i--).  The “.” is at the end so it makes the program faster to do -- instead of ++. Everything is the opposite because we are starting at the end. We then have an if statement for the ASCII values. The ACII value for the “.” is 46. So since the value of the file type is after the “.”, there is a intFT = i + 1. That way the file type is being read. We also have the ASCII value for the “\” which is 92.  We then have the substrings. Substring is when you take only a portion of the string. In myFile, we only need the string which is after the ".". intFT is the location where the "." ends. intFN is the last /.
Ian:  51-85
I did this portion of the code for the File save section of the code. So first the code is going to ask the user to enter the location of where they want to save the file. That location will be saved as the variable saveFilePath. Next the code asks for the user to enter the filename for the file. That will be saved as the variable saveFileName. After that we call the method to get the file name. Now the file is saved so it will read the file from the location you saved. We obviously want this whole process without any user intervention but at our skill level we decided manual entry will be sufficient enough. So we tell the user just that. Then we use an if statement along with a nested for loop to implement a troubleshooting code. Within the if statement we extract the file name and type from readfile using a for loop that does a similar thing to the code from the file load section.  It is basically a continuation of the code.
Domingo: 85-148
The code below allows the program to extract the filename and type of the file we're trying to save. 
After extracting the name and type, the program runs the two through a for loop, recognizing them and later saving both the name and type. 
After saving the two, the program runs through a nested if statement presenting to the user what may have been wrong, i.e wrong path, type or name. 
If all conditions were met properly, the program would notify the user that the file was saved correctly.
Throughout the code for the filesave, the getFileName and saveFile methods are present; however, we used this as a placeholder. The actual code for the two were deemed too complex for us to figure out, so in turn we included temporary code and other placeholders to represent what we wanted it to do.